requirements.txt
numpy 
scipy
pandas
matplotlib
scikit-learn

# ---実験管理・設定---
wandb
omegaconf
hydra-core

# ---PyTorch関連---
torch
torchvision
pytorch-lightning

# models
transformers

# UI/表示
rich

# 追加依存関係
einops
